refactor: Merge cli/deploy tests into a single file
