+++ typegate/tests/e2e/cli/.gitignore
@@ -1,0 +1,1 @@
+/migration.py
