+++ typegate/tests/e2e/cli/select.sh
