--- typegate/tests/e2e/deploy/.gitignore
@@ -1,1 +1,0 @@
-/migration_failure.py
